Add validation for readability/writability.
