+++ enclone_args/src/proc_args2.rs
@@ -3,44 +3,9 @@
 use enclone_core::defs::EncloneControl;
 use io_utils::{open_userfile_for_read, path_exists};
 use rayon::prelude::*;
-use std::fmt::Write;
-use std::fs::{remove_file, File};
 use std::io::BufRead;
 use string_utils::TextUtils;
 use vector_utils::next_diff;
-
-// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
-
-// Test a file for writeability by writing and then deleting it.
-
-pub fn test_writeable(val: &str, evil_eye: bool) -> Result<(), String> {
-    if evil_eye {
-        println!("creating file {val} to test writability");
-    }
-    let f = File::create(val);
-    if f.is_err() {
-        let mut msgx =
-            format!("\nYou've specified an output file\n{val}\nthat cannot be written.\n");
-        if val.contains('/') {
-            let dir = val.rev_before("/");
-            let msg = if path_exists(dir) {
-                "exists"
-            } else {
-                "does not exist"
-            };
-            writeln!(msgx, "Note that the path {dir} {msg}.").unwrap();
-        }
-        return Err(msgx);
-    }
-    if evil_eye {
-        println!("removing file {val}");
-    }
-    remove_file(val).unwrap_or_else(|_| panic!("could not remove file {val}"));
-    if evil_eye {
-        println!("removal of file {val} complete");
-    }
-    Ok(())
-}
 
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
