+++ enclone_args/src/process_special_arg1.rs
@@ -2,13 +2,12 @@
 
 // Process a special argument, i.e. one that does not fit into a neat bucket.
 
-use crate::proc_args2::test_writeable;
 use crate::proc_args2::{is_simple_arg, is_usize_arg};
 use enclone_core::cell_color::{
     CellColor, ColorByCategoricalVariableValue, ColorByDataset, ColorByVariableValue,
 };
 use enclone_core::defs::EncloneControl;
-use enclone_core::tilde_expand_me;
+use enclone_core::{test_writeable, tilde_expand_me};
 use enclone_vars::encode_arith;
 use evalexpr::build_operator_tree;
 use expr_tools::test_functions_in_node;
